refactored/renamed and changed an isEmpty
